Break up some large files and add more comments
